Updated the hash field on the sha256_message struct to 32 bytes only (no need for the null terminator); Deleted the message_schedule field from the sha256_base struct; Added the 6 logical functions used in the hash algorithm; Added the sha256_message_digest() function for digesting the message and producing a hash; Added the sha256_message_show_hash() function; Fixed some bugs;
	-Fixed a bug on the sha256_message_create_from_buffer() function: If the message bits length was a multiple of 8, we would end up masking the last byte. That's because we would end up with a AND mask of 00000000 for bit_length%8 == 0.
	-Updated the hash field on the sha256_message struct to be only 32 bytes long, since it's not a string and we don't need a null terminator. For showing the hash we need to print every single byte as a hexadecimal.
	-Deleted the message_schedule field from the sha256_base struct, since we will use a local variable for this on the sha256_message_digest() function.
	-Added the six logical functions used by the sha256 hashing algorithm.
	-Added the sha256_message_digest() function that processes the message hash. Some notes on this function that are pertinent:
First, the processor being used is little endian. So when working with the 32-bit pieces of the message schedule, we need to make them little endian so the arithmetic and logical operators work. After all the digesting, we convert the 32-bit pieces of the hash back to big-endian.
Second, we need to add a if-statement that checks if the system is big-endian, so we avoid this conversion on this case (not implemented yet).
Third, we need to review the function. It's functional, but after experimenting a lot with it we might have some hidden bug. Check it!
Forth, we need to add a if-statement that checks if the message hash was already computed, so we avoid extra processing of already hashed messages.
	-Added the sha256_message_show_hash() function, that will show the hash in hexadecimal notation and the chars that represent those values.