Fixed an expression that could result in UB
	-The expression modified was submited to the standard integer promotions, resulting in an expression
with all elements cast to 'signed int'. Even if the integer was 32-bit, a value of 0xFF shifted 24 bits to
the left would reach the sign bit, which results in undefined behavior. Also, the code didn't account for
implementations where 'int' is smaller than 32-bits. We now explicitly cast the members of the expression
to 'uint32_t' so we have a complete expression of unsigned 32-bits long types;